+++ sdk/storagecache/arm-storagecache/CHANGELOG.md
@@ -1,5 +1,24 @@
 # Release History
     
+## 5.1.0 (2022-03-10)
+    
+**Features**
+
+  - Added operation group AscUsages
+  - Added operation StorageTargetOperations.beginInvalidate
+  - Added operation StorageTargetOperations.beginInvalidateAndWait
+  - Added Interface AscUsagesListNextOptionalParams
+  - Added Interface AscUsagesListOptionalParams
+  - Added Interface ResourceUsage
+  - Added Interface ResourceUsageName
+  - Added Interface ResourceUsagesListResult
+  - Added Interface StorageTargetInvalidateOptionalParams
+  - Added Type Alias AscUsagesListNextResponse
+  - Added Type Alias AscUsagesListResponse
+  - Interface Cache_2 has a new optional parameter zones
+  - Class StorageCacheManagementClient has a new parameter ascUsages
+    
+    
 ## 5.0.0 (2022-01-21)
 
 The package of @azure/arm-storagecache is using our next generation design principles since version 5.0.0, which contains breaking changes.